rename module `attributes` to `attribute`
