+++ examples/5_extract_pngs.rs
@@ -26,7 +26,7 @@ pub fn main() {
 
     for (layer_index, layer) in image.layers.iter().enumerate() {
         let layer_name = layer.attributes.name.as_ref()
-            .map_or(String::from("main_layer"), attributes::Text::to_string);
+            .map_or(String::from("main_layer"), attribute::Text::to_string);
 
         for channel in &layer.channels {
             match &channel.samples {
